// Soe.Angular/TypeScript/Common/Models/ExpenseDTO.ts
import { IExpenseHeadDTO, IExpenseRowDTO, IExpenseRowGridDTO, IFileUploadDTO } from "../../Scripts/TypeLite.Net4";
import { AccountDTO } from "./AccountDTO";
import { SoeEntityState } from "../../Util/CommonEnumerations";

export class ExpenseHeadDTO implements IExpenseHeadDTO {
    accounting: string;
    actorCompanyId: number;
    comment: string;
    created: Date;
    createdBy: string;
    employeeId: number;
    expenseHeadId: number;
    expenseRows: IExpenseRowDTO[];
    modified: Date;
    modifiedBy: string;
    projectId: number;
    start: Date;
    state: SoeEntityState;
    stop: Date;
    timeBlockDateId: number;
}

export class ExpenseRowDTO implements IExpenseRowDTO {
    customerInvoiceNr: string;
    timePeriodId: number;
    accounting: string;
    actorCompanyId: number;
    amount: number;
    amountCurrency: number;
    amountEntCurrency: number;
    amountLedgerCurrency: number;
    comment: string;
    created: Date;
    createdBy: string;
    customerInvoiceId: number;
    customerInvoiceRowAttestStateId: number;
    customerInvoiceRowId: number;
    employeeId: number;
    expenseHeadId: number;
    expenseRowId: number;
    externalComment: string;
    files: IFileUploadDTO[];
    hasFiles: boolean;
    invoicedAmount: number;
    invoicedAmountCurrency: number;
    invoicedAmountEntCurrency: number;
    invoicedAmountLedgerCurrency: number;
    invoiceRowAttestStateColor: string;
    invoiceRowAttestStateId: number;
    invoiceRowAttestStateName: string;
    isDeleted: boolean;
    isReadOnly: boolean;
    isSpecifiedUnitPrice: boolean;
    isTimeReadOnly: boolean;
    modified: Date;
    modifiedBy: string;
    payrollAttestStateColor: string;
    payrollAttestStateId: number;
    payrollAttestStateName: string;
    projectId: number;
    quantity: number;
    standOnDate: Date;
    start: Date;
    stop: Date;
    timeCodeId: number;
    timeCodeName: string;
    timePayrollAttestStateId: number;
    transferToOrder: boolean;
    unitPrice: number;
    unitPriceCurrency: number;
    unitPriceEntCurrency: number;
    unitPriceLedgerCurrency: number;
    vat: number;
    vatCurrency: number;
    vatEntCurrency: number;
    vatLedgerCurrency: number;       
}

export class ExpenseRowGridDTO implements IExpenseRowGridDTO {
    actorCustomerId: number;
    amount: number;
    amountCurrency: number;
    amountExVat: number;
    comment: string;
    customerName: string;
    employeeId: number;
    employeeName: string;
    employeeNumber: string;
    expenseHeadId: number;
    expenseRowId: number;
    externalComment: string;
    from: Date;
    hasFiles: boolean;
    invoicedAmount: number;
    invoicedAmountCurrency: number;
    invoiceRowAttestStateColor: string;
    invoiceRowAttestStateId: number;
    invoiceRowAttestStateName: string;
    isSpecifiedUnitPrice: boolean;
    orderId: number;
    orderNr: string;
    payrollAttestStateColor: string;
    payrollAttestStateId: number;
    payrollAttestStateName: string;
    payrollTransactionDate: Date;
    projectId: number;
    projectName: string;
    projectNr: string;
    quantity: number;
    timeCodeId: number;
    timeCodeName: string;
    timeCodeRegistrationType: number;
    timePayrollTransactionIds: number[];
    unitPrice: number;
    vat: number;
    vatCurrency: number;
}